chore: move a test between modules
